added MeshInstancePanel
